+++ lib/ooxml_parser/common_parser/common_data/color.rb
@@ -258,22 +258,6 @@ module OoxmlParser
         end
       end
 
-      def parse_color_tag(node)
-        return if node.nil?
-        node.attributes.each do |key, value|
-          case key
-          when 'val'
-            return Color.new.parse_hex_string(value.value.to_s)
-          when 'theme'
-            return ThemeColors.parse_color_theme(node)
-          when 'rgb'
-            return Color.new.parse_hex_string(value.value.to_s)
-          when 'indexed'
-            return Color.get_rgb_by_color_index(value.value.to_i)
-          end
-        end
-      end
-
       def parse_scheme_color(scheme_color_node)
         color = ThemeColors.list[scheme_color_node.attribute('val').value.to_sym]
         hls = HSLColor.rgb_to_hsl(color)